Remove postprocessing from npm, add to git workflow, iGEM-deploy

+++ utils/postprocess.py
@@ -10,16 +10,17 @@ for item in soup.contents:
     if isinstance(item, Doctype):
         item.extract()
 
-with open('dist/index.bundle.js', 'r') as file:
-    script_file = file.read()
+if (os.path.isfile('dist/index.bundle.js')):
+    with open('dist/index.bundle.js', 'r') as file:
+        script_file = file.read()
 
-# embed JS in HTML
-script_tag = soup.find('script')
-script_tag.string = script_file
-del script_tag.attrs['src']
+    # embed JS in HTML
+    script_tag = soup.find('script')
+    script_tag.string = script_file
+    del script_tag.attrs['src']
+
+    os.remove('dist/index.bundle.js')
 
 with open('dist/Poster/index.html', 'wb') as file:
     file.write(b'{{ Poster }}\n\n')
     file.write(soup.prettify('utf-8'))
-
-os.remove('dist/index.bundle.js')
